修改客户和用户管理

+++ src/components/purchase/customer/list.vue
@@ -192,16 +192,20 @@
           return false;
         }
 
-        // 遮罩
-        this.loading = true;
-
-        this.$common.deleteAxios(that, '/purchase/customer/delete', {'id': that.selectedRow}, '用户删除成功').then(function (flag) {
-          if (flag) {
-            that.$common.tableSearch(that, that.url, that.formData);
-            that.selectedRow = '';
-          }
-          that.loading = false;
-        });
+        this.$confirm('是否确认删除数据？<br/><b>数据删除后无法恢复</b>', '提示', {
+          type: 'warning',
+          dangerouslyUseHTMLString: true
+        }).then(function () {
+          // 遮罩
+          this.loading = true;
+          this.$common.deleteAxios(that, '/purchase/customer/delete', {'id': that.selectedRow}, '用户删除成功').then(function (flag) {
+            if (flag) {
+              that.$common.tableSearch(that, that.url, that.formData);
+              that.selectedRow = '';
+            }
+            that.loading = false;
+          });
+        })
       },
       exportBtn() {
         let that = this;
